+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed USB/DFU hub not rebooting after flashing firmware on Windows (pybricksdev#112).
+
+[pybricksdev#112]: https://github.com/pybricks/pybricksdev/issues/112
+
 ## [2.1.0] - 2025-09-10
 
 ### Added